% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_due_date.R
\name{calculate_due_date}
\alias{calculate_due_date}
\title{Calculate a Due Date}
\usage{
calculate_due_date(days)
}
\arguments{
\item{days}{An integer specifying the number of days to add to today's date.}
}
\value{
A `Date` object representing the calculated due date.
}
\description{
This function calculates a due date by adding a specified number of days to today's date.
}
\examples{
# Calculate a due date 10 days from today
calculate_due_date(10)

# Calculate a due date 0 days from today (i.e., today)
calculate_due_date(0)

}
