% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time_diff.R
\name{calculate_time_diff}
\alias{calculate_time_diff}
\title{Calculate Time Difference}
\usage{
calculate_time_diff(target_date, current_time = Sys.time())
}
\arguments{
\item{target_date}{A date input (character or Date) representing the target date.}
}
\value{
A `difftime` object representing the difference in seconds between the target date and the current time.
}
\description{
Computes the time difference in seconds between the current system time and a target date.
}
\examples{
calculate_time_diff(Sys.Date() + 1)  # Difference between now and tomorrow

}
