% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_drinks_data.R
\name{create_drinks_data}
\alias{create_drinks_data}
\title{Utility function to create the drinks data frame}
\usage{
create_drinks_data(input)
}
\arguments{
\item{input}{A list of user inputs containing the quantities for each drink type.}
}
\value{
A data frame with two columns: 'Drink' (name of the drink) and 'Count' (number of drinks).
}
\description{
This function takes the input values from the UI and creates a data frame
with the drink types and their corresponding counts.
}
\examples{
# Assuming input contains the selected quantities
# create_drinks_data(list(coffee = 2, energy_drink = 3, tea = 1, soft_drink = 5, water = 7))
# This returns a data frame with drinks and counts.

}
