% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_clock_server.R
\name{date_clock_server}
\alias{date_clock_server}
\title{Date Clock Server Module}
\usage{
date_clock_server(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\value{
None (side effects: updates countdown timer in the UI).
}
\description{
Handles the countdown logic and updates the remaining time every second.
}
