% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_clock_ui.R
\name{date_clock_ui}
\alias{date_clock_ui}
\title{Date and Countdown Module UI}
\usage{
date_clock_ui(id)
}
\arguments{
\item{id}{A unique identifier for the UI module. This ID is used to namespace the input and output elements.}
}
\value{
A `tagList` containing UI elements, including:
  \itemize{
    \item A text output for displaying the current date and time.
    \item A date input for selecting the target date for the countdown.
    \item A text output for displaying the countdown timer.
  }
}
\description{
This function creates the user interface (UI) for a countdown timer module. It displays the current date and time, allows the user to select a target date for the countdown, and shows the remaining time until the target date.
}
\details{
The UI includes:
  \itemize{
    \item A header and text output to display the current date and time.
    \item A date input field for the user to choose a target date (default is 7 days from the current date).
    \item A countdown timer that updates to show the remaining time until the target date.
  }
}
\examples{
# In a Shiny app, this UI module can be used as follows:
# ui <- fluidPage(
#   date_clock_ui("countdown_module")
# )
# server <- function(input, output, session) {
#   date_clock_server("countdown_module", input, output, session)
# }

}
\seealso{
\code{\link{date_clock_server}} for the corresponding server-side logic.
}
