% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drink_module_ui.R
\name{drink_module_ui}
\alias{drink_module_ui}
\title{Drink Module UI: User Input for Drink Selection and Plot Display}
\usage{
drink_module_ui(id)
}
\arguments{
\item{id}{A unique identifier for the UI module. This ID is used to
namespace the input and output elements.}
}
\value{
A `tagList` containing:
  \itemize{
    \item Numeric inputs for entering the number of each drink type (coffee, energy drinks, tea, soft drinks, water).
    \item A plot output for displaying the barplot of drink quantities.
  }
}
\description{
This function creates a user interface (UI) module for selecting the number
of various drinks (coffee, energy drinks, tea, soft drinks, and water).
It includes input fields for each drink type and renders a plot output
to visualize the selection in the form of a bar plot.
}
\examples{
# In a Shiny app, this UI module can be used as follows:
# ui <- fluidPage(
#   drink_module_ui("drink_selection")
# )

}
