% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius_function.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{A numeric value representing the temperature in Fahrenheit.}
}
\value{
The corresponding temperature in Celsius.
}
\description{
This function converts a temperature from Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32)  # Returns 0
fahrenheit_to_celsius(100) # Returns 37.78

}
