% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius_ui.R
\name{fahrenheit_to_celsius_ui}
\alias{fahrenheit_to_celsius_ui}
\title{Fahrenheit to Celsius UI Module}
\usage{
fahrenheit_to_celsius_ui(id)
}
\arguments{
\item{id}{\code{character} - The module ID. This is used to namespace the UI elements and prevent conflicts.}
}
\value{
A `tagList` containing:
  \itemize{
    \item A numeric input for entering the temperature in fahrenheit.
    \item A text output for displaying the calculated temperature in celsius.
  }
}
\description{
This function creates a Shiny UI module for converting Fahrenheit temperatures to Celsius.
It consists of a numeric input for entering a temperature in Fahrenheit and a text output
to display the equivalent temperature in Celsius.
}
\examples{
In a Shiny app, this UI module can be used as follows:
# ui <- fluidPage(
#    fahrenheit_to_celsius_ui("temp_converter")
# )
# server <- function(input, output, session) {
#   fahrenheit_to_celsius_server("temp_converter", input, output, session)
# }

}
