% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_clock_server.R
\name{date_clock_server}
\alias{date_clock_server}
\title{Countdown Module Server: Server Logic for Date and Countdown Timer}
\usage{
date_clock_server(input, output, session)
}
\arguments{
\item{input}{A list-like object containing the input values from the UI. This includes the `target_date`
input element, which is the target date selected by the user.}

\item{output}{A list-like object containing the output renderers. This includes `current_datetime` for
displaying the current date and time and `countdown` for showing the countdown timer.}

\item{session}{The session object passed from the Shiny app's server function. It is used to manage the app's session state.}
}
\value{
This function does not return a value. It updates the `output$current_datetime` with the current date and time
and the `output$countdown` with the remaining time until the target date.
}
\description{
Defines the server-side logic for displaying the current date and time, as well as a countdown timer
that shows the remaining time until the selected target date. The countdown is updated every second.
}
\details{
The server module:
  \itemize{
    \item Displays the current date and time.
    \item Calculates and displays the countdown to a target date selected by the user.
    \item Updates the countdown every second.
  }
}
\examples{
# In a Shiny app, this server module can be used as follows:
# server <- function(input, output, session) {
#   date_clock_server(input, output, session)
# }

}
\seealso{
\code{\link{date_clock_ui}} for the corresponding UI module.
}
