% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_clock_ui.R
\name{date_clock_ui}
\alias{date_clock_ui}
\title{Date Clock UI Module}
\usage{
date_clock_ui(id)
}
\arguments{
\item{id}{A unique module identifier.}
}
\value{
A `tagList` containing:
- A date input for selecting the target date.
- A text output displaying the countdown timer.
}
\description{
This module provides the UI for a countdown timer where the user selects a target date.
}
