% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius_server.R
\name{fahrenheit_to_celsius_server}
\alias{fahrenheit_to_celsius_server}
\title{Fahrenheit to Celsius Server}
\usage{
fahrenheit_to_celsius_server(id, fahrenheit)
}
\arguments{
\item{id}{A unique identifier for the server module. This ID is used to associate the inputs and outputs
within the module.}

\item{fahrenheit}{A numeric value representing the temperature in Fahrenheit. This parameter is passed
to the module from the parent server or UI.}
}
\value{
This function does not return a value. It updates the `output$celsius` with the converted temperature in Celsius.
}
\description{
This function defines the server-side logic for converting a temperature value from Fahrenheit to Celsius.
It reacts to a `fahrenheit` input and updates the `celsius` output with the converted temperature in Celsius.
}
\details{
The server module calculates the Celsius temperature using the `fahrenheit_to_celsius` function and displays the result
  as a text output.
}
\examples{
# In a Shiny app, this server module can be used as follows:
# server <- function(input, output, session) {
#   fahrenheit_to_celsius_server("fahrenheit_to_celsius", fahrenheit = reactive(input$fahrenheit))
# }

}
\seealso{
\code{\link{fahrenheit_to_celsius_ui}} for the corresponding UI module.
}
